T8511 Split `/budinkite-veikti` into tabs/subpages
