+++ src/components/Header/Header.js
@@ -53,7 +53,7 @@ const Header = () => {
                   <Link to="/protesto-formos/renginiai/">Renginiai</Link>
                 </li>
                 <li>
-                  <Link to="/protesto-formos/budinkite-veikti/">
+                  <Link to="/protesto-formos/budinkite-veikti/ambasada/">
                     Budinkite veikti
                   </Link>
                 </li>
