+++ src/pages/protesto-formos/budinkite-veikti/imone.js
@@ -1,0 +1,113 @@
+import * as React from "react";
+import PropTypes from "prop-types";
+import { graphql } from "gatsby";
+
+import BudinkiteVeikti from "../../../shared-pages/BudinkiteVeikti";
+
+const Page = ({ data }) => {
+  const content = data.contents.edges.map((edge) => {
+    return {
+      ...edge.node.childMarkdownRemark.frontmatter,
+      html: edge.node.childMarkdownRemark.html,
+      excerpt: edge.node.childMarkdownRemark.excerpt,
+    };
+  })[0];
+
+  const addressees = data.addressees.edges.map((edge) => {
+    return edge.node.childMarkdownRemark.frontmatter;
+  });
+
+  return (
+    <BudinkiteVeikti
+      addressees={addressees}
+      content={content}
+      pagePath="/protesto-formos/budinkite-veikti/imone/"
+    />
+  );
+};
+
+export const query = graphql`
+  query {
+    contents: allFile(
+      filter: {
+        sourceInstanceName: { eq: "page-contents" }
+        absolutePath: {
+          regex: "//src/content/pages/protesto-formos/budinkite-veikti.md$/"
+        }
+      }
+    ) {
+      edges {
+        node {
+          childMarkdownRemark {
+            frontmatter {
+              title
+            }
+            html
+            excerpt(format: PLAIN, pruneLength: 160)
+          }
+        }
+      }
+    }
+    addressees: allFile(
+      filter: {
+        sourceInstanceName: { eq: "addressees" }
+        childMarkdownRemark: { frontmatter: { type: { eq: "įmonė" } } }
+      }
+
+      sort: { fields: childMarkdownRemark___frontmatter___weight }
+    ) {
+      edges {
+        node {
+          childMarkdownRemark {
+            frontmatter {
+              type
+              title
+              purpose
+              emailOrLink
+              address
+            }
+          }
+        }
+      }
+    }
+  }
+`;
+
+Page.propTypes = {
+  data: PropTypes.shape({
+    contents: PropTypes.shape({
+      edges: PropTypes.arrayOf(
+        PropTypes.shape({
+          node: PropTypes.shape({
+            childMarkdownRemark: PropTypes.shape({
+              frontmatter: PropTypes.shape({
+                title: PropTypes.string,
+              }),
+            }),
+            html: PropTypes.string,
+            excerpt: PropTypes.string,
+          }),
+        })
+      ),
+    }),
+    addressees: PropTypes.shape({
+      edges: PropTypes.arrayOf(
+        PropTypes.shape({
+          node: PropTypes.shape({
+            childMarkdownRemark: PropTypes.shape({
+              frontmatter: PropTypes.shape({
+                type: PropTypes.array,
+                title: PropTypes.string,
+                purpose: PropTypes.string,
+                emailOrLink: PropTypes.string,
+                address: PropTypes.string,
+              }),
+            }),
+          }),
+        })
+      ),
+    }),
+  }),
+};
+
+export default Page;
