+++ src/pages/protesto-formos/budinkite-veikti/index.js
@@ -1,0 +1,11 @@
+import { useEffect } from "react";
+import { navigate } from "@reach/router";
+
+const Page = () => {
+  useEffect(() => {
+    navigate(`/protesto-formos/budinkite-veikti/ambasada/`);
+  }, []);
+  return null;
+};
+
+export default Page;
